Palmer Dabbelt's Portage Prefix
This prefix is my personal playpen, so stuff here will probably be
marked stable without a large amount of testing aside from on my own
machines.  I'll try and keep the history as clean as possible so it's
easy for people to read.

Feel free to submit patches via email.