Update sys-devel/pconfigure to 0.4.1
I've also removed the old pconfigure ebuilds, as they're not really
useful anymore (I doubt many of them would even build properly).  This
ebuild is the first pconfigure ebuild that doesn't need some sort of
hackery in order to build, which was part of the reason for the cruft
removal (getting rid of files/).