Add epatch_user to app-text/mupdf
I've got a custom patch that I want applied to mupdf, but I can't do
that because epatch_user doesn't appear to be called from
src_prepare().

This just adds epatch_user to the end of every mupdf ebuild.  I'm not
sure if this is the right way to go about this, but it appears to at
least be somewhat working...